+++ src/boost_histogram/accumulators.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function
 
-from ._core.accumulators import Sum, Mean, WeightedSum, WeightedMean
+from ._core.accumulators import Mean, Sum, WeightedMean, WeightedSum
 
 del absolute_import, division, print_function