+++ tests/test_axis_options.py
@@ -1,30 +1,30 @@
 # -*- coding: utf-8 -*-
-from boost_histogram.axis import options
+from boost_histogram.axis import Traits
 
 
 def test_compare():
-    assert options() == options()
-    assert options(underflow=True) == options(underflow=True)
-    assert options(overflow=True) == options(overflow=True)
-    assert options(circular=True) == options(circular=True)
-    assert options(growth=True) == options(growth=True)
-    assert options(underflow=True, overflow=True) == options(
+    assert Traits() == Traits()
+    assert Traits(underflow=True) == Traits(underflow=True)
+    assert Traits(overflow=True) == Traits(overflow=True)
+    assert Traits(circular=True) == Traits(circular=True)
+    assert Traits(growth=True) == Traits(growth=True)
+    assert Traits(underflow=True, overflow=True) == Traits(
         underflow=True, overflow=True
     )
-    assert options() != options(overflow=True)
-    assert options(underflow=True) != options(overflow=True)
-    assert options(underflow=True, overflow=True) != options(overflow=True)
+    assert Traits() != Traits(overflow=True)
+    assert Traits(underflow=True) != Traits(overflow=True)
+    assert Traits(underflow=True, overflow=True) != Traits(overflow=True)
 
 
 def test_fields():
-    o = options()
+    o = Traits()
 
     assert not o.underflow
     assert not o.overflow
     assert not o.circular
     assert not o.growth
 
-    o = options(underflow=True, overflow=True)
+    o = Traits(underflow=True, overflow=True)
 
     assert o.underflow
     assert o.overflow