+++ tests/test_benchmark_1d.py
@@ -1,10 +1,9 @@
 # -*- coding: utf-8 -*-
+import numpy as np
 import pytest
+from numpy.testing import assert_allclose, assert_array_equal
 
 import boost_histogram as bh
-
-import numpy as np
-from numpy.testing import assert_array_equal, assert_allclose
 
 STORAGES = (bh.storage.Int64, bh.storage.Double, bh.storage.Unlimited)
 DTYPES = (np.float64, np.float32, np.int64, np.int32)